Un-pub ColorTable struct
