+++ tests/chessboard-8px-24bit.rs
@@ -20,11 +20,6 @@ fn chessboard_8px_24bit() {
         }
     );
 
-    assert!(
-        bmp.color_table().is_none(),
-        "there should be no color table for this image"
-    );
-
     assert_eq!(bmp.image_data().len(), 314 - 122);
 }
 
@@ -46,11 +41,6 @@ fn chessboard_8px_24bit_truncated_iter() {
             channel_masks: None,
             row_order: RowOrder::BottomUp,
         }
-    );
-
-    assert!(
-        bmp.color_table().is_none(),
-        "there should be no color table for this image"
     );
 
     let pixels: Vec<u32> = bmp.pixels().map(|pixel| pixel.color).collect();
